part-1-agent/pyproject.toml
[project]
name = "part-1-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.97.0",
    "python-dotenv>=1.2.2",
]
